[ADDED] Client side validation of the form #26
The compilation of the form cannot leave empty fields and must respect
the agreements defined with specific regexp for user name and password.
If the compilation is not correct, a specific message is printed in the
modal that appears to cover the form. The sequence of error messages
first concerns the fields left empty and then the regularity first of
the user name and then of the password.